Moved dataset-related classes into org.cirdles.topsoil.data
