+++ src/main/java/org/cirdles/topsoil/app/TopsoilMainWindow.java
@@ -47,9 +47,9 @@ import javafx.stage.FileChooser;
 import javafx.stage.Stage;
 import javafx.stage.Window;
 import org.cirdles.javafx.CustomVBox;
-import org.cirdles.topsoil.app.table.DataSet;
-import org.cirdles.topsoil.app.table.Record;
-import org.cirdles.topsoil.app.table.TSVDataSet;
+import org.cirdles.topsoil.data.DataSet;
+import org.cirdles.topsoil.data.Record;
+import org.cirdles.topsoil.data.TSVDataSet;
 import org.cirdles.topsoil.app.utils.GetApplicationDirectoryOperation;
 import org.cirdles.topsoil.app.utils.GetDocumentsDirectoryOperation;
 import org.cirdles.topsoil.app.utils.TSVTableReader;
