+++ src/main/java/org/cirdles/topsoil/data/DataSet.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.app.table;
+package org.cirdles.topsoil.data;
 
 import java.nio.file.Path;
