+++ src/main/java/org/cirdles/topsoil/data/TSVDataSet.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.app.table;
+package org.cirdles.topsoil.data;
 
 import java.io.IOException;
 import java.nio.file.Files;
